python/text-to-speech/requirements.txt
boto3==1.28.9
azure-cognitiveservices-speech==1.24.0
google-cloud-texttospeech==2.14.1
parameterized==0.9.0
